# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Login",
                "description": "Login",
                "operationId": "Login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Enter User Email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "user@yopmail.com"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Enter Password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "123456"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                }
            }
        },
        "/api/v1/signup": {
            "post": {
                "tags": [
                    "Signup"
                ],
                "summary": "Signup",
                "description": "User Signup",
                "operationId": "Signup",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Enter User Name(Only alphabets, Special characters not allowed and Space allowed)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Kartik"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Enter User Email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "user@yopmail.com"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}